Fixed certain bugs in Google sync

+++ CEO/src/cs2103/TaskList.java
@@ -26,6 +26,7 @@ public class TaskList {
 	private GoogleEngine google;
 	private ArrayList<Task> tasks;
 	private static final String SYNCING = "Syncing with Google, please wait for a while";
+	private static final String SYNCING_ERROR = "Some error occurred when commit changes to Google";
 	
 	private TaskList(Option option) throws FatalException, HandledException{
 		this.dataFile = new File("CEOStore.ics");
@@ -137,7 +138,7 @@ public class TaskList {
 					this.google.addTask(task);
 				}
 			} catch (IOException e) {
-				this.google = null;
+				CommonUtil.print(SYNCING_ERROR);
 			}
 		}
 		return this.getTaskByTask(task);
@@ -158,7 +159,7 @@ public class TaskList {
 					}
 				}
 			} catch (IOException e) {
-				this.google = null;
+				CommonUtil.print(SYNCING_ERROR);
 			}
 		}
 		return this.getTaskByTask(task);
@@ -175,7 +176,7 @@ public class TaskList {
 					this.google.deleteTask(task);
 				}
 			} catch (IOException e) {
-				this.google = null;
+				CommonUtil.print(SYNCING_ERROR);
 			}
 		}
 	}
@@ -189,6 +190,7 @@ public class TaskList {
 			syncToGoogle(googleList);
 			this.tasks = this.storage.getTaskList();
 		} catch (FatalException | HandledException | IOException e) {
+			e.printStackTrace();
 			this.google = null;
 		}
 	}
@@ -197,11 +199,17 @@ public class TaskList {
 		for (Task remoteTask:googleList){
 			Task localTask = this.getTaskByTask(remoteTask, this.tasks);
 			if (localTask == null){
-				this.storage.updateTask(remoteTask);
-			} else {
-				if (localTask.getLastModified().before(remoteTask.getLastModified())){
-					remoteTask.updateCreated(localTask.getCreated());
+				if (!remoteTask.isDeleted()){
 					this.storage.updateTask(remoteTask);
+				}
+			} else {
+				if (remoteTask.isDeleted()){
+					this.storage.deleteTask(remoteTask);
+				} else {
+					if (localTask.getLastModified().before(remoteTask.getLastModified())){
+						remoteTask.updateCreated(localTask.getCreated());
+						this.storage.updateTask(remoteTask);
+					}
 				}
 			}
 		}
@@ -218,7 +226,9 @@ public class TaskList {
 				}
 			} else {
 				if (localTask.isDeleted()){
-					this.google.deleteTask(remoteTask);
+					if (!remoteTask.isDeleted()){
+						this.google.deleteTask(remoteTask);
+					}
 				} else {
 					if (remoteTask.getLastModified().before(localTask.getLastModified())){
 						this.google.updateTask(localTask);
